Move callback handlers to their own directory
